web development - 20170819

+++ src/main/java/com/stemcloud/smart/web/dao/SensorCameraRepository.java
@@ -11,5 +11,5 @@ import java.util.List;
  * Description:
  */
 public interface SensorCameraRepository extends CrudRepository<SensorCamera, Long> {
-    List<SensorCamera> findBySensorId(int sensorId);
+    List<SensorCamera> findBySensorId(long sensorId);
 }
